+++ vendor_lib/rspec/support.rb
@@ -1,6 +1,0 @@
-require "rspec/support/version"
-
-module RSpec
-  module Support
-  end
-end