+++ CHANGELOG.md
@@ -1,4 +1,12 @@
-## edge
+## 0.5.0beta2
+*   This is the RSpec31asyncsubject branch
+*   By default, any subject, it example block, before(:each), after(:each), and around that returns a promise will be executed asynchronously. Async is NOT yet supported for context level hooks.
+
+## 0.5.0beta1
+
+*   Update to RSpec 3.1 (core is 3.1.7, expectations/support 3.1.2, mocks 3.1.3)
+
+*   Remove copy of source (and just rely on git submodule fetch)
 
 *   Remove predicate matcher fixes as Opal supports $1..$9 special gvars.